requirements.txt
kafka-python==2.3.0
pyspark
pymongo
pandas
neo4j